backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
pandas
nba_api
python-dotenv
